Fix dissapearing less button closes #42
The show less button dissapears after click show more for the second time. It looks like it was originally planned that there would be a seperate behaviour for explicit and implicit expand and collapse but I can't see that an implicit expand and collapse was ever implemented. As such all open and closes are explicit and the show less button shouldn't be removed unless we want to remove it (in which case it can be done manually in the appication logic)